backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
psycopg2-binary==2.9.9
supabase==2.15.3